requirements.txt
amqp==2.1.4
billiard==3.5.0.2
celery==4.0.2
certifi==2017.4.17
chardet==3.0.4
coreapi==2.3.1
coreschema==0.0.4
defusedxml==0.5.0
dj-database-url==0.4.2
Django==1.11.2
django-cas-ng==3.5.8
django-common-helpers==0.9.1
django-cors-headers==2.1.0
django-crontab==0.7.1
django-filter==1.0.4
django-smartfields==1.0.9
djangorestframework==3.6.3
djangorestframework-filters==0.10.1
djangorestframework-jwt==1.10.0
drfdocs==0.0.11
gunicorn==19.7.1
idna==2.5
itypes==1.1.0
Jinja2==2.9.6
kombu==4.0.2
Markdown==2.6.8
MarkupSafe==1.0
mongoengine==0.13.0
oauthlib==2.0.2
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.1
Pygments==2.2.0
PyJWT==1.5.0
pymongo==3.4.0
python-cas==1.2.0
python3-openid==3.1.0
pytz==2017.2
requests==2.18.1
requests-oauthlib==0.8.0
setupfiles==0.0.50
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
vine==1.1.3
whitenoise==3.3.0
